backend/requirements.txt
Flask
Flask-Cors
keras
numpy
pillow
